requirements.txt
numpy
opencv-python~=4.4.0.44
Pillow~=7.2.0